[feat] 최대값 개선 (#84)
* fix - Update maximum months limit to 11989(999 year) and add tests for investment calculations

* test - Refactor savings product balance calculation tests to use parameterized tests

* test - Refactor exception handling in FinancialProduct tests for better readability

* fix - Update AnnualInterestRate to use BigDecimal and adjust validation for maximum rate

* fix - Refactor tax rate handling to use FixedTaxRate and update related calculations

* fix - Rename tax rate retrieval methods to use getValue for consistency

* feat - rename field, rate to value

* fix - Update tax rate methods to return BigDecimal instead of double for consistency

* fix - Update tax rate methods to return BigDecimal instead of double for consistency

* fix - Update FixedTaxRate to use mutable value field and add JPA annotations

* fix - Update investment amount handling to use BigDecimal for precision

* fix - Update maximum months limit to 11989(999 year) and add tests for investment calculations

* fix - Update AnnualInterestRate to use BigDecimal and adjust validation for maximum rate

* fix - Refactor tax rate handling to use FixedTaxRate and update related calculations

* fix - Rename tax rate retrieval methods to use getValue for consistency

* feat - rename field, rate to value

* fix - Update tax rate methods to return BigDecimal instead of double for consistency

* fix - Update tax rate methods to return BigDecimal instead of double for consistency

* fix - Update FixedTaxRate to use mutable value field and add JPA annotations

* fix - Update investment amount handling to use BigDecimal for precision

* fix - Update taxRate handling to use FixedTaxRate for consistency across financial products

* fix - change TaxRate to FixedTaxRate

* fix - delete doubleValue

* fix - update hashCode method to handle trailing zeros and add equals test for FixedTaxRate

* fix - implement custom equals and hashCode methods for ProductAmount and ProductMonths classes

* fix AnnualInterestRate

* fix comment

* fix error message desc

* fix - update hashCode method to return 0 for null values and add test for ProductAmount hashCode equality

* fix - test display name

* fix - replace ProductRate with AnnualInterestRate in financial product classes

* fix - add JPA column annotation to AnnualInterestRate value field

+++ src/main/java/co/invest72/investment/domain/investment/CompoundFixedDeposit.java
@@ -167,7 +167,7 @@ public class CompoundFixedDeposit implements Investment {
 	}
 
 	@Override
-	public double getTaxRate() {
+	public BigDecimal getTaxRate() {
 		return taxable.getTaxRate();
 	}
 }
