+++ src/main/java/co/invest72/investment/domain/investment/SimpleFixedDeposit.java
@@ -167,7 +167,7 @@ public class SimpleFixedDeposit implements Investment {
 	}
 
 	@Override
-	public double getTaxRate() {
+	public BigDecimal getTaxRate() {
 		return taxable.getTaxRate();
 	}
 }
